Split auth.go into login and logout
